[Build] Fix latest MSVC and CLang compiler errors (#1008)
* Explicit cast specification

* Explicit cast specification

* Suppress -Wno-tautological-compare in Clang 10 and above

https://reviews.llvm.org/rG8b0d14a8f0cc085afa2a9c86c237da81c74517fc

* Explicit cast specification

* Add HAVE_STDINT_H compiler flag

* Explicit casting mechanism type specification

* typo

+++ amxmodx/amxmodx.cpp
@@ -1063,7 +1063,7 @@ static cell AMX_NATIVE_CALL show_menu(AMX *amx, cell *params) /* 3 param */
 				pPlayer->vgui = false;
 				
 				if (time == -1)
-					pPlayer->menuexpire = INFINITE;
+					pPlayer->menuexpire = static_cast<float>(INFINITE);
 				else
 					pPlayer->menuexpire = gpGlobals->time + static_cast<float>(time);
 				
@@ -1090,7 +1090,7 @@ static cell AMX_NATIVE_CALL show_menu(AMX *amx, cell *params) /* 3 param */
 			pPlayer->vgui = false;			
 
 			if (time == -1)
-				pPlayer->menuexpire = INFINITE;
+				pPlayer->menuexpire = static_cast<float>(INFINITE);
 			else
 				pPlayer->menuexpire = gpGlobals->time + static_cast<float>(time);
